Move HeaderState and SnackbarState to the global reducer
