+++ src/views/login/login.view.tsx
@@ -6,8 +6,8 @@ import * as globalActions from "src/store/global/global.actions";
 import * as loginActions from "src/store/login/login.actions";
 import * as loginThunks from "src/store/login/login.thunks";
 import { Step } from "src/store/login/login.reducer";
+import { HeaderState } from "src/store/global/global.reducer";
 import { AsyncTask, isAsyncTaskDataAvailable } from "src/utils/types";
-import { HeaderState } from "src/views/app.view";
 import WalletButtonList from "src/views/login/components/wallet-button-list/wallet-button-list.view";
 import WalletLoader from "src/views/login/components/wallet-loader/wallet-loader.view";
 import CreateAccountAuth from "src/views/login/components/create-account-auth/create-account-auth.view";
