+++ src/views/transactions/deposit/deposit.view.tsx
@@ -9,10 +9,10 @@ import * as depositThunks from "src/store/transactions/deposit/deposit.thunks";
 import * as depositActions from "src/store/transactions/deposit/deposit.actions";
 import * as globalThunks from "src/store/global/global.thunks";
 import { changeHeader } from "src/store/global/global.actions";
+import { HeaderState } from "src/store/global/global.reducer";
 import useTransactionStyles from "src/views/transactions/deposit/deposit.styles";
 import TransactionOverview from "src/views/transactions/components/transaction-overview/transaction-overview.view";
 import AccountSelector from "src/views/transactions/components/account-selector/account-selector.view";
-import { HeaderState } from "src/views/app.view";
 import Spinner from "src/views/shared/spinner/spinner.view";
 import DepositForm from "src/views/transactions/deposit/components/deposit-form/deposit-form.view";
 import * as storage from "src/utils/storage";
